Update to UI
This is mainly an update to the GUI and the functionallity on the UI experience. This update is mainly to make the program eaiser to run and was just kinda something fun to do to pass some time. Added some new packages that are needed for the program that need to be install so make sure to run the requirements file again to get all the packages it needs.