Update
Small updates to media so that all of the information is being read in to the program as sanitized hex. Then in the writer I changed the writing to be for binary files for media. This seems to correctly work for all of the media layer. I will test the jpg saving this way and see if it will work. Pushing this for now as a good saftey spot and start of dev branch.